ProgressBar and CrossCorrelator updates.
CrossCorrelator has a check on the event number of the last zoomed in
correlations to check work isn't repeated.
ProgressBar now has a signal handler, which it uses when the new inc
function is used to increment the internal counter. The ++ operator
is now deprecated. When SIGINT is received, it sets the reference
variable entry to numEntries. Pretty pleased with that.