Updates to CrossCorrelator and RootTools.
Widened the 380 notch in RootTools as some events in the minBias sample
aren't entirely filtered by it.
Added a min bias selection function to RootTools to have it in just one
place. Also turned off the PPS2 (G12) selection as it was only on around
MCM and just slows down my reco scripts.