Restore next, prev, first, last, play, reverse button functionality
Basic functionality of the main 6 buttons is restored. They work
playing through the events and with a cut applied (the applyCut
functionality has also been restored. The code base still prints
annoying error messages with the main body of the old code commented
out.

Selecting an event by event number is still not working and there is
replacement for the eventNumber index functionality.