MagicControlPanel doesn't segfault on quitting.
There's some kind of blinky timer inside TGTextEntryField that makes
the cursor flash. If you delete the window while that cursor is
present on the screen all kinds of horrible things happen. The
solution is to set the focus away from the TGTextEntryFields before
closing the window. I think that one eluded evan Ryan...