Better FourierBuffer integration with Interferometry mode.
Press L to load history into the FourierBuffer so you can examine the
event exactly how it appears in analysis scripts.

Press P for the probability metric, press A for the rayleigh
amplitudes, press C for the reduced chisquare, and press R for the
regular reconstruction.

Had to add an optional force boolian to getEventViewerCanvas to allow
the event to be redrawn after loading history.

Passed Adu5Pat pointer to AnalysisReco since Cosmin added that
feature.

Happy MagicDisplaying!